Panel drawing & layout displayed
Have now made a main method to create a panel and draw the layout of the
different parts of the GUI, have adjusted variable in MiniMap so it fits
with the template (roughly). Still need to get better positioning for
room on display and make it isometric. Also need to get the listener
object made and passing correctly.